@@ -3,7 +3,7 @@
 # Automatically generated by nengo-bones, do not edit this file directly
 
 import io
-import os
+import pathlib
 import runpy
 
 try:
@@ -26,8 +26,8 @@ def read(*filenames, **kwargs):
     return sep.join(buf)
 
 
-root = os.path.dirname(os.path.realpath(__file__))
-version = runpy.run_path(os.path.join(root, "nengo_fpga", "version.py"))["version"]
+root = pathlib.Path(__file__).parent
+version = runpy.run_path(str(root / "nengo_fpga" / "version.py"))["version"]
 
 install_req = [
     "nengo>=3.0.0",
@@ -44,6 +44,7 @@ docs_req = [
 ]
 optional_req = []
 tests_req = [
+    "nengo[tests]>=3.0.0",
     "pytest>=3.6",
     "pytest-mock>=2.0",
     "pytest-cov>=2.6",
@@ -68,7 +69,7 @@ setup(
         "optional": optional_req,
         "tests": tests_req,
     },
-    python_requires=">=3.5",
+    python_requires=">=3.6",
     package_data={
         "nengo_fpga": [
             "fpga_config",
