@@ -2,7 +2,7 @@
 #
 # Automatically generated by nengo-bones, do not edit this file directly
 
-import os
+import pathlib
 
 import nengo_fpga
 
@@ -85,11 +85,11 @@ release = nengo_fpga.__version__  # Full version, with tags
 templates_path = ["_templates"]
 html_static_path = ["_static"]
 html_theme = "nengo_sphinx_theme"
-html_title = "NengoFPGA {0} docs".format(release)
+html_title = f"NengoFPGA {release} docs"
 htmlhelp_basename = "NengoFPGA"
 html_last_updated_fmt = ""  # Default output format (suppressed)
 html_show_sphinx = False
-html_favicon = os.path.join("_static", "favicon.ico")
+html_favicon = str(pathlib.Path("_static", "favicon.ico"))
 html_theme_options = {
     "nengo_logo": "nengo-fpga-full-light.svg",
     "nengo_logo_color": "#541a8b",
